requirements.txt
mysql-connector-python==8.3.0
Faker
pyshark
pandas
tqdm